requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
autogen-agentchat==0.4.9.1
autogen-core==0.4.9.1
autogen-ext==0.4.9.1
certifi==2025.1.31
charset-normalizer==3.4.1
deprecated==1.2.18
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==8.6.1
jiter==0.9.0
jsonref==1.1.0
openai==1.66.3
opentelemetry-api==1.31.0
pillow==11.1.0
protobuf==5.29.3
pydantic==2.10.6
pydantic-core==2.27.2
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
tiktoken==0.9.0
tqdm==4.67.1
typing-extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
zipp==3.21.0
streamlit
jinja2
python-dotenv
supabase
google-generativeai
pyautogen
